requirements.txt
aiosqlite
python-telegram-bot==20.5
telethon==1.32.1
openai==1.8.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.24.1
schedule==1.2.0 